requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-environ==0.4.5
djangorestframework==3.12.4
filelock==3.0.12
idna==3.2
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.1
tldextract==3.1.0
urllib3==1.26.6
